flask-backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
convex==0.6.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-health-checking==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.24.6
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langsmith==0.1.121
llama-cloud==0.0.15
llama-index==0.11.5
llama-index-agent-openai==0.3.0
llama-index-cli==0.3.0
llama-index-core==0.11.5
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-llama-cpp==0.2.1
llama-index-llms-openai==0.2.2
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.2
llama_cpp_python==0.2.90
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
minijinja==2.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.43.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.0.1
setuptools==75.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
striprtf==0.0.26
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.34.0
weaviate-client>=3.26.7,<4.0.0
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.9.11
